setup data prop
