+++ examples/src/index.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import './index.css';
-import App from './../../src/App';
+import NetworkViewer from './../../src/NetworkViewer';
+import networkData from './../../tests/__fixtures__/network.json';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(<NetworkViewer data={networkData} />, document.getElementById('root'));
