+++ src/Containers/FilterContainer.jsx
@@ -5,7 +5,7 @@ import {
 
 import ImportHar from './../Components/ImportHAR';
 import Search from './../Components/Filters/Search';
-import { useNetwork } from './../state/network/provider';
+import { useNetwork } from './../state/network/Context';
 import { FILTERS } from './../constants';
 import Styles from './FilterContainer.styles';
