+++ webpack.config.js
@@ -20,7 +20,7 @@ module.exports = {
       test: /\.css$/,
       use: ['style-loader', 'css-loader'],
     }, {
-      test: /\.module\.s(a|c)ss$/,
+      test: /\.styles\.s(a|c)ss$/,
       loader: [
         isDevelopment ? 'style-loader' : MiniCssExtractPlugin.loader,
         {
